requirements/requirements.txt
pydantic>=1.8.2
tensorflow>=2.5.1
numpy==1.19.5
requests>=2.25,<2.30
twine==3.3.0
PyYAML~=5.4.1
h5py
pymongo~=3.11.3
backoff~=1.10.0
click~=7.1.2
pandas~=1.2.3
matplotlib~=3.3.4
python-jose[cryptography]~=3.3.0
boto3~=1.17.39
tensorflow_privacy@git+https://github.com/tensorflow/privacy.git@331a2911ac3cdf1d9d8d301813be3541257d5bf4#egg=tensorflow_privacy
azure-functions~=1.7.0
